quantization/requirements.txt
autoawq>=0.1.7
gguf>=0.6.0
protobuf
sentencepiece
